优化项目工时与工时详情关联，excel表头映射，数据类型调整，删除无用的类

+++ src/main/java/cn/com/tarotframework/server/oak/dto/ProjectHourDetail.java
@@ -28,7 +28,7 @@ public class ProjectHourDetail {
     private LocalDate fillDate;
 
     // 使用工时
-    private BigDecimal useHour;
+    private Double useHour;
 
     // 项目状态：a 进行中，b运维，c 结束
     private String projectStatus;
